backend/requirements.txt
fastapi==0.115.6
uvicorn==0.33.0
pydantic-settings==2.7.0
python-dotenv==1.0.1
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.159.0
python-jose[cryptography]==3.3.0
